Generates a flextable Object with Error Message

Takes a vector of messages and returns a flextable object with the text in the table. This can be used in automated table generation to cascade an error message to the end user.

Usage

mk_error_tab(msgs)

Arguments

msgs

Vector of error messages

Value

list with a single flextable object

Examples

error_tab = mk_error_tab("This is an error")
error_tab$ft[[1]]

Error

This is an error