Reads and Checks report_fit.yaml File

Reads in the report yaml file and looks it to make sure it has all the necessary fields for the given report.

yaml_read_fit(obnd = NULL, rptyaml = NULL, ph = NULL, fit = NULL)

Arguments

obnd

onbrand report object to have report elements appended to

rptyaml

yaml file containing the report elements and structure

ph

Placeholders in the yaml file can be overwritten at runtime with a named list for example RUN may be "RUNN" in the yaml file. To overwrite this just provide list(RUN="RUN_1") (default: NULL)

fit

nlmixr2 fit object to be reported

Value

List containing the following information about the report

"isgood" - Boolean variable indicating success or failure

"msgs" - Vector of messages

"rpttype" - Type of onbrand report ("Word" or "PowerPoint")

"rptfigfmt" - Default figure formatting (orientation and dimensions)

"rpttabfmt" - Default table formatting (orientation and dimensions)

"rptdetails" - Contents of the yaml file

"rptcont" - Contents of the report to generate